Allow for stringified option keys and stringified "via" option
When config caches come from YAML files and the like it's sometimes difficult to trace whether a specific config key is a symbol. This leads to obscure error messages. Therefore it's sane to recursively symbolize the option hash keys and do a little allowance for the delivery option flag to be a string. This makes Pony friends with YAML configs without extraneous to_sym calls.