Allow for stringified keys in configs
For example, hashes from YAML configs are usually string keyed, however we do not want to convert everything we receive into symbols since this can pollute the symbol table and lead to memory leaks. We therefore use a proxy class which acts sort of like the HashWithIndifferentAccess proxy, and will fetch is values from the hash with string keys when requested by symbol. The downside is that we need to wrap calls to options with this decorator every time, which we do using a delegate.